Break up the entrypoint
